build: Exclude netsim-next crates from CMake build
The bluetooth and rootcanal rust crates are for netsim-next and intended
to be built exclusively with Bazel.

This change removes them from the CMake build configuration
rust/CMakeLists.txt and rust/Cargo.toml to prevent them from being
compiled during the CMake build process. This simplifies the CMake build
by excluding components that are not yet integrated with it.

Bug: 449998651
Test: scripts/build_tools.py
Change-Id: I0d3a774232532a6e5b9150c6b2283eb22f6e12d2