Add script to build and upload cargo doc for netsim-next
This change add a `build-docs.sh` script to generate cargo doc for the `netsim-next` crate and optionally upload it to x20 Betosim team folder.

The script does `cargo doc` to build the documentation and `cargo-modules` to generate a module dependency graph, which is now included in the `netsim-next` crate's documentation.

Bug: 433973644
Test: build-docs.sh
Change-Id: I3f34196daf968fa856116d388cb22493ae3f0a16